+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- New linter errors.
+### Changed
+- Use ErrStack instead Error in state.
+- Error's workflow of state.
 
 ## [0.3.4] - 2019-09-17
 ### Fixed